Fix Subscribe form UX
