+++ app/dashboard/src/components/AriaComponents/Inputs/variants.ts
@@ -76,73 +76,3 @@ export const INPUT_STYLES = tv({
     variant: 'outline',
   },
 })
-
-export const SELECTOR_STYLES = tv({
-  base: 'block w-full overflow-hidden bg-transparent transition-[border-color,outline] duration-200',
-  variants: {
-    disabled: {
-      true: { base: 'cursor-default opacity-50', textArea: 'cursor-default' },
-      false: { base: 'cursor-text', textArea: 'cursor-text' },
-    },
-    readOnly: { true: 'cursor-default' },
-    size: {
-      medium: { base: '' },
-    },
-    rounded: {
-      none: 'rounded-none',
-      small: 'rounded-sm',
-      medium: 'rounded-md',
-      large: 'rounded-lg',
-      xlarge: 'rounded-xl',
-      xxlarge: 'rounded-2xl',
-      xxxlarge: 'rounded-3xl',
-      full: 'rounded-full',
-    },
-    variant: {
-      outline: {
-        base: 'border-[0.5px] outline-offset-2 border-primary/20 focus-within:outline focus-within:outline-2 focus-within:outline-offset-[-1px] focus-within:outline-primary focus-within:border-primary/50',
-        textArea: 'border-transparent focus-within:border-transparent',
-      },
-    },
-  },
-  defaultVariants: {
-    size: 'medium',
-    rounded: 'xlarge',
-    variant: 'outline',
-  },
-  slots: {
-    radioGroup: 'flex',
-  },
-})
-
-export const SELECTOR_OPTION_STYLES = tv({
-  base: TEXT_STYLE({
-    className: 'flex flex-1 items-center min-h-6 relative overflow-auto',
-    variant: 'body',
-  }),
-  variants: {
-    rounded: {
-      none: 'first:rounded-l-none last:rounded-r-none',
-      small: 'first:rounded-l-sm last:rounded-r-sm',
-      medium: 'first:rounded-l-md last:rounded-r-md',
-      large: 'first:rounded-l-lg last:rounded-r-lg',
-      xlarge: 'first:rounded-l-xl last:rounded-r-xl',
-      xxlarge: 'first:rounded-l-2xl last:rounded-r-2xl',
-      xxxlarge: 'first:rounded-l-3xl last:rounded-r-3xl',
-      full: 'first:rounded-l-full last:rounded-r-full',
-    },
-    size: {
-      medium: { base: 'px-[11px] pb-1.5 pt-2' },
-      small: { base: 'px-[11px] pb-0.5 pt-1' },
-    },
-    variant: {
-      primary:
-        'selected:bg-primary selected:text-white not-selected:hover:bg-primary/20 border-r-[0.5px] border-primary/20',
-    },
-  },
-  defaultVariants: {
-    size: 'medium',
-    rounded: 'xlarge',
-    variant: 'primary',
-  },
-})
