+++ app/ide-desktop/common/src/text/english.json
@@ -535,6 +535,7 @@
   "originalPrice": "Original Price",
   "youSave": "You Save",
   "totalPrice": "Total Price",
+  "subtotalPrice": "Subtotal Price",
   "returnToDashboard": "Return to Dashboard",
   "subscribeTitle": "Enso Pricing",
   "subscribeSuccessSubmit": "Go to Dashboard",
